[#14095] Split JsonAppendFunction into distinct functions for String and Array operations
