+++ server/resp/src/main/java/org/infinispan/server/resp/ExternalizerIds.java
@@ -12,10 +12,11 @@ public interface ExternalizerIds {
    Integer JSON_LEN_FUNCTION = 2206;
    Integer JSON_TYPE_FUNCTION = 2207;
    Integer JSON_DEL_FUNCTION = 2208;
-   Integer JSON_APPEND_FUNCTION = 2209;
+   Integer JSON_ARRAY_APPEND_FUNCTION = 2209;
    Integer JSON_TOGGLE_FUNCTION = 2210;
    Integer JSON_OBJKEYS_FUNCTION = 2211;
    Integer JSON_NUMINCRBY_FUNCTION = 2212;
    Integer JSON_ARRINDEX_FUNCTION = 2213;
    Integer JSON_BUCKET = 2214;
+   Integer JSON_STRING_APPEND_FUNCTION = 2215;
 }
